# public/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Storage API",
        "description": "This is documentation for API",
        "contact": {
            "email": "apiteam@wordnik.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "depot.esmc.info",
    "basePath": "/api/v1",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/item_attributes/": {
            "get": {
                "tags": [
                    "Item attribute"
                ],
                "summary": "Return all item attributes with pagination",
                "description": "This is method for find all item attributes",
                "operationId": "findAllItemAttributes",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Number of page",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "title": "data",
                            "type": "object",
                            "$ref": "#/definitions/ItemAttribute"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Item attribute"
                ],
                "summary": "Create new item attributes",
                "description": "This is method create new item attribute",
                "operationId": "storeItemAttribute",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of item attribute",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "alias",
                        "in": "query",
                        "description": "Alias of organization",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type of item attribute",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "integer",
                            "float",
                            "string",
                            "text",
                            "boolean",
                            "datetime"
                        ]
                    },
                    {
                        "name": "unit_id",
                        "in": "query",
                        "description": "ID of unit",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "schema": {
                            "title": "data",
                            "type": "object",
                            "$ref": "#/definitions/ItemAttribute"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/item_attributes/{id}": {
            "get": {
                "tags": [
                    "Item attribute"
                ],
                "summary": "Return one item attribute",
                "description": "This is method for find one item attribute by ID",
                "operationId": "findItemAttribute",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of item",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "title": "data",
                            "type": "object",
                            "$ref": "#/definitions/ItemAttribute"
                        }
                    },
                    "404": {
                        "description": "Item not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Item attribute"
                ],
                "summary": "Update item attribute",
                "description": "This is method update depot organization role by ID",
                "operationId": "updateItemAttribute",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of item attribute",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of item attribute",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "alias",
                        "in": "query",
                        "description": "Alias of organization",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type of item attribute",
                        "required": false,
                        "type": "string",
                        "enum": [
                            "integer",
                            "float",
                            "string",
                            "text",
                            "boolean",
                            "datetime"
                        ]
                    },
                    {
                        "name": "unit_id",
                        "in": "query",
                        "description": "ID of unit",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "schema": {
                            "title": "data",
                            "type": "object",
                            "$ref": "#/definitions/ItemAttribute"
                        }
                    },
                    "404": {
                        "description": "Item not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Item attribute"
                ],
                "summary": "Remove item attributes by ID",
                "description": "This is method remove item attribute",
                "operationId": "destroyItemAttribute",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID item attribute for remove",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "bad request"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ItemAttribute": {
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "name": {
                                "type": "string"
                            },
                            "alias": {
                                "type": "string"
                            },
                            "type": {
                                "type": "string"
                            },
                            "created_at": {
                                "type": "string",
                                "format": "dateTime"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "dateTime"
                            },
                            "deleted_at": {
                                "type": "string",
                                "format": "dateTime"
                            },
                            "unit_id": {
                                "type": "integer"
                            }
                        }
                    }
                }
            },
            "type": "object"
        },
...
    },
...
}